Fix tests for tampered sig and modified payload
Two tests used to modify a single byte of signature or payload by
XORing it with itself, causing the target byte to be zero.

If the original data in the target byte was already zero, then nothing
was changed, causing the test to fail.

Fix this by changing bitwise operation to XOR the target byte with a 1,
so the target byte will always change by 1 bit.

Closes #68