Add `RS256` algorithm constant
The `AlgorithmRS256` was added as a known COSE signing
algorithm.

Attempts to create a new Signer through [NewSigner] will result in
an error, because the `rsaSigner` does not implement PKCS#1 v1.5
signing nor verification. If PKCS#1 v1.5 support is required, one
has to provide an external implementation of a [cose.Signer].

Signed-off-by: hslatman <herman+github@smallstep.com>